Add OpenPGP integration with cross-distribution compatibility and fix compiler warnings
- Implement full OpenPGP compatibility using GnuPG and Nitrokey hardware
- Add comprehensive C API for signing, verification, encryption, decryption
- Support both software-based (GnuPG) and hardware-accelerated (Nitrokey) operations
- Add cross-distribution binary detection for gpg, gpg2, nitropy, gnuradio-companion
- Include distribution-specific installation instructions for Ubuntu, Debian, Fedora, RHEL, Arch, openSUSE, Alpine, Gentoo
- Add comprehensive test suite and documentation
- Fix AX.25 overflow warning by changing uint8_t to uint16_t for max_frame_length
- Fix all write() and fread() warnings with proper error checking and cleanup
- Add OpenSSL KDF header to resolve HKDF function declaration warnings
- Improve error handling throughout OpenPGP integration
- Add example usage demonstrating both GnuPG and Nitrokey operations
- Ensure compatibility across Linux distributions with robust binary detection

+++ libm17/tnc/ax25_protocol.h
@@ -96,7 +96,7 @@ typedef struct {
     uint16_t slot_time;              // Slot time (10ms units)
     uint8_t tx_tail;                 // TX tail (10ms units)
     bool full_duplex;                // Full duplex mode
-    uint8_t max_frame_length;        // Maximum frame length
+    uint16_t max_frame_length;       // Maximum frame length
     uint8_t window_size;             // Window size
     uint32_t t1_timeout;             // T1 timeout (ms)
     uint32_t t2_timeout;             // T2 timeout (ms)
